Initialize project
The idea is to have a multi-module Maven project where each module represents a minimal vulnerable application that can be containerized easily.

We use those for testing rogue-jndi functionality.

The first module is a servlet that's served by Tomcat 10.x. Because it uses a JDK version below 15, a payload using the ScriptEngineManager can be used.